requirements.txt
numpy
pandas
scikit-learn
tensorflow-macos
keras
seaborn
pyarrow
fastparquet
soundfile
torchaudio